requirements.txt
python-dotenv==1.0.1
pandas==2.2.1
redis==5.0.3
sqlalchemy==2.0.28
aiohttp==3.9.3
websockets==12.0